Refactor Preconditions formatting style regex into two checks, one for Guava, one for logsafe
